refactored to examples and updated todos
