+++ pega-prototyping/README.md
@@ -1,10 +1,21 @@
 # Pega Prototyping Experiment
- 
- A demo to see how we can prototype with Pega's constellation design system
- Uses vite to compile React, index.html contains the shell for the SPA
 
- NOTES:
+A demo to see how we can prototype with Pega's constellation design system
+Uses vite to compile React, index.html contains the shell for the SPA
 
- - [ ] SPA by default, do we implement a router for prototyping? Make it easy to configure.
- - [ ] If it proves to be a success, do we try to make it like GDS? 
+TODOS:
 
+- [ ] Update with proper HMRC branded colours and logo svg/png
+- [ ] Chat with Ian/Peter about the idea of this for myself and Alex.
+- [ ] Attempt to recreate AUI
+- [ ] Attempt to recreate Child benefits process page
+- [ ] Bring in Theme Cosmos/UI Kit for co-existence prototypes
+- [ ] Abstract away app shell, with props to customise for future projects
+
+CONSIDERATIONS:
+
+- [ ] SPA by default, do we implement a router for prototyping? Make it easy to configure.
+- [ ] If it proves to be a success, do we try to make it like GDS?
+- [ ] What do we do about onboarding?
+- [ ] Do we publish as an npm package i.e. "hmrc-pega-prototype-kit"?
+- [ ] Prettier extension to avoid manually formatting react code
